Move GenomeBrowserIdsContext in its own folder
